cloud_function/requirements.txt
functions-framework==3.3.0
google-cloud-storage==2.8.0
google-cloud-firestore==2.11.0
google-cloud-aiplatform==1.25.0
vertexai==0.5.0
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.2
langchain-text-splitters==0.0.1
unstructured==0.10.30
unstructured-inference==0.7.21
